Drop the build; ship src/index.js directly on k6
The library is plain JS that k6 runs as-is. The esbuild bundle + yarn/eslint
tooling only existed to produce a single distributable file, which we can do
without a build.

- Inline src/junit.js and src/text.js into a single self-contained ES module
  src/index.js (function bodies unchanged; only the CommonJS exports become ESM
  exports). Same public API: jUnit, textSummary, humanizeValue.
- Remove package.json, yarn.lock, esbuild/eslint/prettier/yarn/npm config and
  .gitignore.
- checks.yml: run the k6 smoke test via grafana/setup-k6-action instead of
  yarn lint/build.
- release.yml: produce build.tar.gz by copying src/index.js to build/index.js
  (same artifact shape jslib.k6.io consumes), no bundling.
- Add .editorconfig and test/smoke.test.js; update README and AGENTS.md.
