--- .eslintrc
@@ -1,25 +1,0 @@
-{
-    "parserOptions": {
-      "ecmaFeatures": {
-        "jsx": true
-      },
-      "ecmaVersion": 2020,
-      "sourceType": "module"
-    },
-    "extends": ["eslint:recommended"],
-    "plugins": ["prettier"],
-    "env": {
-      "browser": true,
-      "node": true
-    },
-    "rules": {
-      "prettier/prettier": ["error"],
-      "arrow-body-style": "warn",
-      "camelcase": 0,
-      "object-curly-newline": 0,
-      "operator-linebreak": 0,
-      "no-shadow": 0,
-      "max-len": [2, 120],
-      "no-underscore-dangle": "off"      
-    }
-  }
