+++ .github/workflows/release.yml
@@ -6,7 +6,7 @@ on:
       - v*
 
 jobs:
-  checks:
+  release:
     runs-on: ubuntu-latest
     if: startsWith(github.ref, 'refs/tags/v')
     permissions:
@@ -15,15 +15,11 @@ jobs:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6
         with:
           persist-credentials: false
-      - name: Use Node.js
-        uses: actions/setup-node@a0853c24544627f65ddf259abe73b1d18a591444 # v5
-        with:
-          node-version: '18.x'
-          package-manager-cache: false
-      - name: Dependencies
-        run: yarn
-      - name: Check build & archive build
-        run: yarn build && tar -zcf build.tar.gz build
+      - name: Archive module
+        run: |
+          mkdir -p build
+          cp src/index.js build/index.js
+          tar -zcf build.tar.gz build
       - name: Create Release & Upload Asset
         env:
             GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
