--- .prettierignore
@@ -1,4 +1,0 @@
-.cache/
-public/
-node_modules/
-build/
