--- .yarnrc.yml
@@ -1,1 +1,0 @@
-enableScripts: false
