--- package.json
@@ -1,25 +1,0 @@
-{
-  "name": "k6-jslib-summary",
-  "repository": "https://github.com/grafana/k6-jslib-summary",
-  "version": "0.0.5",
-  "packageManager": "npm@10.9.2",
-  "description": "Creates a summary of the k6 JS libraries used in a test runs.",
-  "devDependencies": {
-    "esbuild": "0.28.0",
-    "eslint": "8.57.1",
-    "eslint-config-prettier": "10.1.8",
-    "eslint-plugin-prettier": "5.5.5",
-    "prettier": "3.8.3"
-  },
-  "engines": {},
-  "scripts": {
-    "lint": "eslint --max-warnings 0 src",
-    "fix": "eslint --fix src",
-    "build": "esbuild ./src/index.js --bundle --sourcemap --minify --format=cjs --legal-comments=none --outfile=./build/index.js"
-  },
-  "keywords": [
-    "k6"
-  ],
-  "author": "grafana/k6 team",
-  "license": "Apache-2.0"
-}
